Update pc_workload.service
This is a systemd service unit file that defines a service called "PC workload collector service" that will run a Python script called "Combined.py" as a simple process with a user "minmin". 

The [Unit] section provides a human-readable description for the service and specifies the network.target as a requirement for the service to start.

The [Service] section defines the service's behavior. User=minmin specifies the user that will run the service. Type=simple indicates that the service is a simple, long-running process. ExecStart specifies the command to start the service, which in this case is /usr/bin/python3 followed by the path to the Combined.py script. ExecStop specifies the command to stop the service, which in this case sends a QUIT signal to the process ID of the main process of the service. Restart=always indicates that the service will be restarted automatically if it crashes.

The [Install] section specifies when the service should be started. WantedBy=multi-user.target indicates that the service should be started when the system is in multi-user mode.